core: Improve XDVDFSFilesystem operation speed
- Adds scan_dirent_tree function that reads directory entries
  iteratively
- Specializes copy operations in XDVDFSFilesystem to use std::io::copy
  when able
- Replace path walking in read_dir and copy operations with a trie
  lookup and cache

+++ xdvdfs-core/src/util.rs
@@ -4,7 +4,7 @@ use core::fmt::{Debug, Display};
 #[derive(Debug)]
 pub enum Error<E> {
     IOError(E),
-    SizeOutOfBounds(u32, u32),
+    SizeOutOfBounds(u64, u32),
     SerializationFailed(bincode::Error),
     InvalidVolume,
     DirectoryEmpty,
